Refactor tests to not use a Django project
