+++ .travis.yml
@@ -8,4 +8,4 @@ env:
 install:
   - "pip install -r requirements.txt"
   - "pip install Django==$DJANGO_VERSION"
-script: "python src/manage.py test testapp"
+script: "python src/auditlog_tests/runtests.py"
